Move match test to other file
